chore: extend example component
